Struct gas::tournaments::single_elimination::SingleElimination

pub struct SingleElimination<G: Game<N, NSYMS>, const N: usize, const NSYMS: usize> {
    pub game: G,
}

Fields
game: G
Implementations

impl<G: Game<N, NSYMS>, const N: usize, const NSYMS: usize> SingleElimination<G, N, NSYMS>

pub const fn new(game: G) -> SingleElimination<G, N, NSYMS>

pub fn do_side(
&self,
population: &Vec<Candidate<N, NSYMS>>,
remaining: &mut VecDeque<usize>,
rating: &mut Vec<usize>,
k: f64,
rng: &mut Rando,
score_weights: &Vec<f64>
)

the guts of the single elimination tournament, pulled out so it can be called for each side of a double elimination tournament.

remaining: indices into population of the candidates that have not yet lost out of the tournament. By taking 2 candidates from the top and inserting the winner into the bottom, the bye’s for tournament sizes not a power of 2 are handled properly. rating: the ELO rating of candidates. Indices in this vector match that of population k: the K factor for ELO rating. This is reduced by 10% every round of the tournament.